requirements.txt
openai
gradio
tqdm
python-dotenv